+++ SAXPY_0/saxpy_0.go
@@ -10,13 +10,13 @@ import (
 func main() {
 	var n int
 	n,_= strconv.Atoi(os.Args[1])
-	var a float64 = 2
-	x:= make([]float64,n)
-	y:= make([]float64,n)
+	var a float32 = 2
+	x:= make([]float32,n)
+	y:= make([]float32,n)
 	//pragma gomp parallel for
 	for i:=0;i<n;i++{
-		x[i]=float64(rand.Int())
-		y[i]=float64(rand.Int())
+		x[i]=float32(rand.Int())
+		y[i]=float32(rand.Int())
 	}
 	
 	//pragma gomp parallel for
