+++ SAXPY_Bench/saxpy_init.go
@@ -9,14 +9,14 @@ func _init_numCPUs() {
 	runtime.GOMAXPROCS(_numCPUs)
 	}
 
-func Saxpy_init(size int) ([]float64, []float64) {
+func Saxpy_init(size int) ([]float32, []float32) {
 
-	x := make([]float64, size)
-	y := make([]float64, size)
+	x := make([]float32, size)
+	y := make([]float32, size)
 
 	for i := 0; i < size; i++ {
-		x[i] = float64(rand.Int())
-		y[i] = float64(rand.Int())
+		x[i] = float32(rand.Int())
+		y[i] = float32(rand.Int())
 	}
 	return x, y
 }
