+++ SAXPY_Original_Parallel/saxpy_original_parallel.go
@@ -14,9 +14,9 @@ func _init_numCPUs() {
 func main() {
 	_init_numCPUs()
 	var n int = 10
-	var a float64 = 2
-	x := []float64{1, 2, 3, 4, 5, 6, 7, 8, 9, 10}
-	y := []float64{1, 2, 3, 4, 5, 6, 7, 8, 9, 10}
+	var a float32 = 2
+	x := []float32{1, 2, 3, 4, 5, 6, 7, 8, 9, 10}
+	y := []float32{1, 2, 3, 4, 5, 6, 7, 8, 9, 10}
 	fmt.Println("Vector x antes del parallel:", x)
 	fmt.Println("Vector y antes del parallel:", y)
 	var _barrier_0_bool = make(chan bool)
